chore: Refactor dynamic row config dialog (#4446)
* Add unique constraint on invitations to ensure no duplicate invitations get created

* Update DynamicRowDialog to new Event based trigger

* Rename for easier identification

* Fix typscript error

* Use EventKeys

* Move event type back to event hook

* Don't need isCustomEvent

* Revert "Add unique constraint on invitations to ensure no duplicate invitations get created"

This reverts commit 4317aaae1db3326119a164a7181f25f6925f9688.

* remove todo

* Add unique constraint on invitations to ensure no duplicate invitations get created

* cleanup some repetitive calls

* Revert "Add unique constraint on invitations to ensure no duplicate invitations get created"

This reverts commit 4317aaae1db3326119a164a7181f25f6925f9688.
